Move Django back end into a single package
