+++ toy_settings/config.py
@@ -1,9 +1,9 @@
 from __future__ import annotations
 
 from .application.unit_of_work import Committer
+from .django_back_end.queries import DjangoRepo
+from .django_back_end.unit_of_work import DjangoCommitter
 from .domain.queries import Repository
-from .repositories.django.repo import DjangoRepo
-from .units_of_work.django import DjangoCommitter
 
 
 def get_repository() -> Repository:
